added IgnoreUnsupported option for TogNMINotifications method
Sometimes YANG files use unsupported types and therefore the
`TogNMINotifications` method throws an error (e.g., if a `GoStruct` contains
a keyless list) and does not return any notifications. But there could
be cases where someone would like to ignore these errors and instead
would like to get back all notifications which could be processed as a
result.

This first draft adds a additional option for the `TogNMINotifications`
method and allows to ignore some errors.